+++ src/contracts/settlement/PostSettlementSurplusAuctionHouse.sol
@@ -64,7 +64,7 @@ contract PostSettlementSurplusAuctionHouse is Authorizable, Modifiable, IPostSet
     bids[_id].bidAmount = _initialBid;
     bids[_id].amountToSell = _amountToSell;
     bids[_id].highBidder = msg.sender;
-    bids[_id].auctionDeadline = uint48(block.timestamp) + _params.totalAuctionLength;
+    bids[_id].auctionDeadline = block.timestamp + _params.totalAuctionLength;
 
     safeEngine.transferInternalCoins(msg.sender, address(this), _amountToSell);
 
@@ -79,7 +79,7 @@ contract PostSettlementSurplusAuctionHouse is Authorizable, Modifiable, IPostSet
     if (_id == 0 || _id > auctionsStarted) revert PSSAH_AuctionNeverStarted();
     if (bids[_id].auctionDeadline > block.timestamp) revert PSSAH_AuctionNotFinished();
     if (bids[_id].bidExpiry != 0) revert PSSAH_BidAlreadyPlaced();
-    bids[_id].auctionDeadline = uint48(block.timestamp) + _params.totalAuctionLength;
+    bids[_id].auctionDeadline = block.timestamp + _params.totalAuctionLength;
     emit RestartAuction(_id, bids[_id].auctionDeadline);
   }
 
@@ -105,7 +105,7 @@ contract PostSettlementSurplusAuctionHouse is Authorizable, Modifiable, IPostSet
     protocolToken.safeTransferFrom(msg.sender, address(this), _bid - bids[_id].bidAmount);
 
     bids[_id].bidAmount = _bid;
-    bids[_id].bidExpiry = uint48(block.timestamp) + _params.bidDuration;
+    bids[_id].bidExpiry = block.timestamp + _params.bidDuration;
 
     emit IncreaseBidSize(_id, msg.sender, _amountToBuy, _bid, bids[_id].bidExpiry);
   }
@@ -130,8 +130,8 @@ contract PostSettlementSurplusAuctionHouse is Authorizable, Modifiable, IPostSet
     uint256 _uint256 = _data.toUint256();
 
     if (_param == 'bidIncrease') _params.bidIncrease = _uint256;
-    else if (_param == 'bidDuration') _params.bidDuration = uint48(_uint256);
-    else if (_param == 'totalAuctionLength') _params.totalAuctionLength = uint48(_uint256);
+    else if (_param == 'bidDuration') _params.bidDuration = _uint256;
+    else if (_param == 'totalAuctionLength') _params.totalAuctionLength = _uint256;
     else revert UnrecognizedParam();
   }
 }
